feat: :art: Change random color of food at each intake
Change random color of food at each intake  I have updated the game logic so that the color of the food changes randomly each time the snake consumes it. Now, when generating a new food square using the `createRandomFood` function, a random color is chosen from the food color palette and assigned to that square. The `drawSquare` function has been modified to correctly handle the third parameter (color) and reflect the change in the user interface.  This change improves the visual variability of the game, making the experience more dynamic and engaging for the player.